+++ lockfile_generator/src/maven.rs
@@ -7,6 +7,17 @@ use crate::{Error, Generator, Result};
 
 pub struct Maven;
 
+#[cfg(not(windows))]
+fn maven_command() -> Command {
+    Command::new("mvn")
+}
+
+#[cfg(windows)]
+fn maven_command() -> Command {
+    // Maven uses a batch script on Windows
+    Command::new("mvn.cmd")
+}
+
 impl Generator for Maven {
     fn lockfile_path(&self, manifest_path: &Path) -> Result<PathBuf> {
         let project_path = manifest_path
@@ -17,7 +28,7 @@ impl Generator for Maven {
 
     fn command(&self, manifest_path: &Path) -> Command {
         let lockfile_path = self.lockfile_path(manifest_path).unwrap();
-        let mut command = Command::new("mvn");
+        let mut command = maven_command();
         command.args([
             "-q",
             "help:effective-pom",