+++ CHANGELOG.md
@@ -19,6 +19,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Fixed
 
 - Lockfile parsing of bun-generated yarn lockfiles
+- Maven lockfile generation on Windows
 
 ### Removed